# docs/swagger.yaml
basePath: /
definitions:
...
  domain.AddressUpdateDTO:
    properties:
      city:
        type: string
      complement:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      state:
        type: string
      street:
        type: string
    type: object
...
  domain.GetProfileResponseDTO:
    properties:
      address:
        $ref: '#/definitions/domain.ProfileAddressDTO'
      biography:
        type: string
      categories:
        items:
          $ref: '#/definitions/domain.ProfileCategoryDTO'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      reviews:
        items: {}
        type: array
      services:
        items:
          $ref: '#/definitions/domain.ServiceResponseDTO'
        type: array
      user_type:
        type: string
    type: object
...
  domain.ProfileAddressDTO:
    properties:
      city:
        type: string
      complement:
        type: string
      neighborhood:
        type: string
      number:
        type: string
      state:
        type: string
      street:
        type: string
    type: object
  domain.ProfileCategoryDTO:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  domain.ServiceCategoryDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  domain.ServiceResponseDTO:
    properties:
      actuation_days:
        items:
          type: string
        type: array
      category:
        $ref: '#/definitions/domain.ServiceCategoryDTO'
      description:
        type: string
      end_time:
        type: string
      id:
        type: string
      name:
        type: string
      offer_since:
        type: string
      photos:
        items:
          type: string
        type: array
      start_time:
        type: string
      value:
        type: number
    type: object
  domain.UpdateProfileRequestDTO:
    properties:
      address:
        $ref: '#/definitions/domain.AddressUpdateDTO'
      biography:
        type: string
      categories:
        items:
          type: integer
        type: array
      email:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the Helptrix helper marketplace platform.
  title: Helptrix API
  version: "1.0"
paths:
...
  /user/profile/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a user profile. Only the owner can delete.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user profile
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Returns a user profile by ID, with optional filters for business
        users
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Filter services by category ID (business only)
        in: query
        name: category_id
        type: integer
      - description: Filter services by actuation day (business only)
        in: query
        name: actuation_day
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetProfileResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Updates mutable fields of a user profile. Only the owner can update.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update profile request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequestDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"